+++ lib/models/src/contracts_module.freezed.dart
@@ -505,3 +505,176 @@ abstract class _LoginResponse implements LoginResponse {
   _$$LoginResponseImplCopyWith<_$LoginResponseImpl> get copyWith =>
       throw _privateConstructorUsedError;
 }
+
+ChangePasswordResponse _$ChangePasswordResponseFromJson(
+    Map<String, dynamic> json) {
+  return _ChangePasswordResponse.fromJson(json);
+}
+
+/// @nodoc
+mixin _$ChangePasswordResponse {
+  bool get flag => throw _privateConstructorUsedError;
+  String get message => throw _privateConstructorUsedError;
+
+  /// Serializes this ChangePasswordResponse to a JSON map.
+  Map<String, dynamic> toJson() => throw _privateConstructorUsedError;
+
+  /// Create a copy of ChangePasswordResponse
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  $ChangePasswordResponseCopyWith<ChangePasswordResponse> get copyWith =>
+      throw _privateConstructorUsedError;
+}
+
+/// @nodoc
+abstract class $ChangePasswordResponseCopyWith<$Res> {
+  factory $ChangePasswordResponseCopyWith(ChangePasswordResponse value,
+          $Res Function(ChangePasswordResponse) then) =
+      _$ChangePasswordResponseCopyWithImpl<$Res, ChangePasswordResponse>;
+  @useResult
+  $Res call({bool flag, String message});
+}
+
+/// @nodoc
+class _$ChangePasswordResponseCopyWithImpl<$Res,
+        $Val extends ChangePasswordResponse>
+    implements $ChangePasswordResponseCopyWith<$Res> {
+  _$ChangePasswordResponseCopyWithImpl(this._value, this._then);
+
+  // ignore: unused_field
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
+
+  /// Create a copy of ChangePasswordResponse
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? flag = null,
+    Object? message = null,
+  }) {
+    return _then(_value.copyWith(
+      flag: null == flag
+          ? _value.flag
+          : flag // ignore: cast_nullable_to_non_nullable
+              as bool,
+      message: null == message
+          ? _value.message
+          : message // ignore: cast_nullable_to_non_nullable
+              as String,
+    ) as $Val);
+  }
+}
+
+/// @nodoc
+abstract class _$$ChangePasswordResponseImplCopyWith<$Res>
+    implements $ChangePasswordResponseCopyWith<$Res> {
+  factory _$$ChangePasswordResponseImplCopyWith(
+          _$ChangePasswordResponseImpl value,
+          $Res Function(_$ChangePasswordResponseImpl) then) =
+      __$$ChangePasswordResponseImplCopyWithImpl<$Res>;
+  @override
+  @useResult
+  $Res call({bool flag, String message});
+}
+
+/// @nodoc
+class __$$ChangePasswordResponseImplCopyWithImpl<$Res>
+    extends _$ChangePasswordResponseCopyWithImpl<$Res,
+        _$ChangePasswordResponseImpl>
+    implements _$$ChangePasswordResponseImplCopyWith<$Res> {
+  __$$ChangePasswordResponseImplCopyWithImpl(
+      _$ChangePasswordResponseImpl _value,
+      $Res Function(_$ChangePasswordResponseImpl) _then)
+      : super(_value, _then);
+
+  /// Create a copy of ChangePasswordResponse
+  /// with the given fields replaced by the non-null parameter values.
+  @pragma('vm:prefer-inline')
+  @override
+  $Res call({
+    Object? flag = null,
+    Object? message = null,
+  }) {
+    return _then(_$ChangePasswordResponseImpl(
+      flag: null == flag
+          ? _value.flag
+          : flag // ignore: cast_nullable_to_non_nullable
+              as bool,
+      message: null == message
+          ? _value.message
+          : message // ignore: cast_nullable_to_non_nullable
+              as String,
+    ));
+  }
+}
+
+/// @nodoc
+@JsonSerializable()
+class _$ChangePasswordResponseImpl implements _ChangePasswordResponse {
+  _$ChangePasswordResponseImpl({required this.flag, required this.message});
+
+  factory _$ChangePasswordResponseImpl.fromJson(Map<String, dynamic> json) =>
+      _$$ChangePasswordResponseImplFromJson(json);
+
+  @override
+  final bool flag;
+  @override
+  final String message;
+
+  @override
+  String toString() {
+    return 'ChangePasswordResponse(flag: $flag, message: $message)';
+  }
+
+  @override
+  bool operator ==(Object other) {
+    return identical(this, other) ||
+        (other.runtimeType == runtimeType &&
+            other is _$ChangePasswordResponseImpl &&
+            (identical(other.flag, flag) || other.flag == flag) &&
+            (identical(other.message, message) || other.message == message));
+  }
+
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  int get hashCode => Object.hash(runtimeType, flag, message);
+
+  /// Create a copy of ChangePasswordResponse
+  /// with the given fields replaced by the non-null parameter values.
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  @override
+  @pragma('vm:prefer-inline')
+  _$$ChangePasswordResponseImplCopyWith<_$ChangePasswordResponseImpl>
+      get copyWith => __$$ChangePasswordResponseImplCopyWithImpl<
+          _$ChangePasswordResponseImpl>(this, _$identity);
+
+  @override
+  Map<String, dynamic> toJson() {
+    return _$$ChangePasswordResponseImplToJson(
+      this,
+    );
+  }
+}
+
+abstract class _ChangePasswordResponse implements ChangePasswordResponse {
+  factory _ChangePasswordResponse(
+      {required final bool flag,
+      required final String message}) = _$ChangePasswordResponseImpl;
+
+  factory _ChangePasswordResponse.fromJson(Map<String, dynamic> json) =
+      _$ChangePasswordResponseImpl.fromJson;
+
+  @override
+  bool get flag;
+  @override
+  String get message;
+
+  /// Create a copy of ChangePasswordResponse
+  /// with the given fields replaced by the non-null parameter values.
+  @override
+  @JsonKey(includeFromJson: false, includeToJson: false)
+  _$$ChangePasswordResponseImplCopyWith<_$ChangePasswordResponseImpl>
+      get copyWith => throw _privateConstructorUsedError;
+}