+++ lib/services/src/api_client.g.dart
@@ -304,13 +304,14 @@ class _ApiClient implements ApiClient {
   }
 
   @override
-  Future<void> changePassword(Map<String, dynamic> data) async {
+  Future<ChangePasswordResponse> changePassword(
+      Map<String, dynamic> data) async {
     final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
     _data.addAll(data);
-    final _options = _setStreamType<void>(Options(
+    final _options = _setStreamType<ChangePasswordResponse>(Options(
       method: 'POST',
       headers: _headers,
       extra: _extra,
@@ -326,17 +327,26 @@ class _ApiClient implements ApiClient {
           _dio.options.baseUrl,
           baseUrl,
         )));
-    await _dio.fetch<void>(_options);
+    final _result = await _dio.fetch<Map<String, dynamic>>(_options);
+    late ChangePasswordResponse _value;
+    try {
+      _value = await compute(deserializeChangePasswordResponse, _result.data!);
+    } on Object catch (e, s) {
+      errorLogger?.logError(e, s, _options);
+      rethrow;
+    }
+    return _value;
   }
 
   @override
-  Future<void> changePasswordMobile(Map<String, dynamic> data) async {
+  Future<ChangePasswordResponse> changePasswordMobile(
+      Map<String, dynamic> data) async {
     final _extra = <String, dynamic>{};
     final queryParameters = <String, dynamic>{};
     final _headers = <String, dynamic>{};
     final _data = <String, dynamic>{};
     _data.addAll(data);
-    final _options = _setStreamType<void>(Options(
+    final _options = _setStreamType<ChangePasswordResponse>(Options(
       method: 'POST',
       headers: _headers,
       extra: _extra,
@@ -352,7 +362,15 @@ class _ApiClient implements ApiClient {
           _dio.options.baseUrl,
           baseUrl,
         )));
-    await _dio.fetch<void>(_options);
+    final _result = await _dio.fetch<Map<String, dynamic>>(_options);
+    late ChangePasswordResponse _value;
+    try {
+      _value = await compute(deserializeChangePasswordResponse, _result.data!);
+    } on Object catch (e, s) {
+      errorLogger?.logError(e, s, _options);
+      rethrow;
+    }
+    return _value;
   }
 
   @override